convert all decimals to integers for conversion results from postgresql

+++ backend/db/migrations/20240921194238-create-spot.js
@@ -34,11 +34,11 @@ module.exports = {
           allowNull: false,
         },
         lat: {
-          type: Sequelize.DECIMAL,
+          type: Sequelize.INTEGER,
           allowNull: false,
         },
         lng: {
-          type: Sequelize.DECIMAL,
+          type: Sequelize.INTEGER,
           allowNull: false,
         },
         name: {
@@ -50,7 +50,7 @@ module.exports = {
           allowNull: false,
         },
         price: {
-          type: Sequelize.DECIMAL,
+          type: Sequelize.INTEGER,
           allowNull: false,
         },
         createdAt: {
